requirements.txt
motor~=3.3.1
python-dotenv~=1.0.0
